Use urssh on remote to get info
Rather than sending all the functions every time, just use the urssh
script on the remote computer to get the data. This is done through a
new, internal action called "get" that gets info. The volume of the
script (quiet/normal/verbose) is passed to the other one.

This also allows for more customizable commands, which can now be given
by the user through the --command command line option.

Also add more output when being verbose, and make sure messages turned
on by verbose go to stderr, not stdout.